changed the Compare_Output_CLI.py and the IBD_compare_output_classes.py to the IBD_compare_output_full.py so that it would not use up so much memory. It seems to work now without using 300 Gb of ram.
