+++ .gitignore
@@ -10,3 +10,4 @@ script_to_understand/
 *.zip
 8_18_20/
 8_19_20/
+exm652227_pairs/
